# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Numerology CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List all rules
      parameters:
        - name: rule_type
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /rules/{id}:
    get:
      operationId: getRule
      tags: [rules]
      summary: Get a rule by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          description: Not found
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rules/{id}/toggle:
    patch:
      operationId: toggleRule
      tags: [rules]
      summary: Toggle rule active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
...
components:
  schemas:
...
    Rule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rule_type:
          type: string
        description:
          type: ["string", "null"]
        condition_json:
          type: object
          additionalProperties: true
        result_json:
          type: object
          additionalProperties: true
        priority:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - rule_type
        - condition_json
        - result_json
        - priority
        - is_active
        - created_at
        - updated_at
    RuleInput:
      type: object
      properties:
        name:
          type: string
        rule_type:
          type: string
        description:
          type: string
        condition_json:
          type: object
          additionalProperties: true
        result_json:
          type: object
          additionalProperties: true
        priority:
          type: integer
        is_active:
          type: boolean
      required:
        - name
        - rule_type
        - condition_json
        - result_json
    RuleUpdate:
      type: object
      properties:
        name:
          type: string
        rule_type:
          type: string
        description:
          type: string
        condition_json:
          type: object
          additionalProperties: true
        result_json:
          type: object
          additionalProperties: true
        priority:
          type: integer
        is_active:
          type: boolean
...