# lib/api-zod/src/generated/types/rule.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */
import type { RuleConditionJson } from './ruleConditionJson';
import type { RuleResultJson } from './ruleResultJson';

export interface Rule {
  id: number;
  name: string;
  rule_type: string;
  /** @nullable */
  description?: string | null;
  condition_json: RuleConditionJson;
  result_json: RuleResultJson;
  priority: number;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}